refactor: 불필요한 컴포넌트화 private 적용
