+++ app/src/main/java/nextstep/signup/ui/auth/screen/SignUpScreen.kt
@@ -3,11 +3,16 @@ package nextstep.signup.ui.auth.screen
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -20,14 +25,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import nextstep.signup.R
-import nextstep.signup.ui.auth.component.SignUpConfirmButton
 import nextstep.signup.ui.auth.component.SignUpForm
-import nextstep.signup.ui.auth.component.SignUpTitle
 import nextstep.signup.ui.auth.model.SignUpFormState
 import nextstep.signup.ui.auth.preview.SignUpPreviewParamsProvider
 import nextstep.signup.ui.interaction.clearFocusWith
@@ -117,6 +121,53 @@ internal fun SignUpScreen(
         SignUpConfirmButton(onDoneSignUp, enableSignUp)
     }
 }
+
+@Composable
+private fun SignUpTitle() {
+    Text(
+        text = stringResource(id = R.string.sign_up_title),
+        style = MaterialTheme.typography.titleLarge.copy(
+            fontWeight = FontWeight.Bold
+        )
+    )
+}
+
+@Composable
+private fun SignUpConfirmButton(onClickSignUp: () -> Unit, enableSignUp: Boolean) {
+    Button(
+        modifier = Modifier
+            .fillMaxWidth(),
+        onClick = { onClickSignUp() },
+        enabled = enableSignUp
+    ) {
+        Text(
+            modifier = Modifier.padding(vertical = 6.dp),
+            text = stringResource(id = R.string.sign_up_confirm_button),
+            style = MaterialTheme.typography.bodyMedium
+        )
+    }
+}
+
+@Preview(name = "enabled")
+@Composable
+private fun PreviewSignUpConfirmButton() {
+    SignupTheme {
+        Surface {
+            SignUpConfirmButton({}, true)
+        }
+    }
+}
+
+@Preview(name = "disabled")
+@Composable
+private fun PreviewSignUpConfirmButton2() {
+    SignupTheme {
+        Surface {
+            SignUpConfirmButton({}, false)
+        }
+    }
+}
+
 
 @Preview(showSystemUi = true)
 @Composable
